requirements.txt
tensorflow
tensorflow-hub
numpy
matplotlib
gradio
Pillow